social-publishing/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
alembic==1.14.0
APScheduler==3.10.4
requests==2.32.3
cryptography==43.0.3
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.21.1
anthropic==0.42.0
gunicorn==23.0.0
python-magic==0.4.27
